requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
click==8.1.8
dnspython==2.7.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
motor==3.6.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.0.3
pydantic_core==2.27.2
pymongo==4.9.2
redis==5.2.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-jose==3.3.0
pyotp==2.9.0
PyJWT==2.8.0
asyncpg==0.30.0
bcrypt==3.2.0
alembic==1.14.0
httpx==0.28.1
google-generativeai==0.8.4
python-dotenv==1.0.1
jinja2
motor==3.6.0
weasyprint==64.0
